iOS 10.0

+++ Frameworks/AVFoundation.framework/AVAsset.h
@@ -8,13 +8,13 @@
 
 @property (getter=MP_canAffectNetworkPlayability, setter=MP_setCanAffectNetworkPlayability:, nonatomic) BOOL MP_canAffectNetworkPlayability;
 @property (readonly) NSArray *availableChapterLocales;
-@property (nonatomic, readonly) struct { int x1; int x2; unsigned int x3; int x4; } duration;
+@property (nonatomic, readonly) struct { long long x1; int x2; unsigned int x3; long long x4; } duration;
 @property (nonatomic, readonly) BOOL isProxy;
-@property (nonatomic, readonly) struct CGSize { double x1; double x2; } naturalSize;
-@property (nonatomic, readonly) struct CGSize { double x1; double x2; } naturalSizeWithPreferredTransforms;
-@property (nonatomic, readonly) double preferredRate;
-@property (nonatomic, readonly) struct CGAffineTransform { double x1; double x2; double x3; double x4; double x5; double x6; } preferredTransform;
-@property (nonatomic, readonly) double preferredVolume;
+@property (nonatomic, readonly) struct CGSize { float x1; float x2; } naturalSize;
+@property (nonatomic, readonly) struct CGSize { float x1; float x2; } naturalSizeWithPreferredTransforms;
+@property (nonatomic, readonly) float preferredRate;
+@property (nonatomic, readonly) struct CGAffineTransform { float x1; float x2; float x3; float x4; float x5; float x6; } preferredTransform;
+@property (nonatomic, readonly) float preferredVolume;
 @property (nonatomic, readonly) id propertyListForProxy;
 @property (nonatomic, readonly) NSValue *pu_cachedDuration;
 @property (setter=rc_setComposedAVURL:, nonatomic, retain) NSURL *rc_composedAVURL;
@@ -30,7 +30,7 @@
 - (id)_URLSessionDataDelegate;
 - (id)_URLSessionOperationQueue;
 - (id)_absoluteURL;
-- (unsigned int)_addChapterMetadataItem:(id)arg1 timeRange:(struct { struct { int x_1_1_1; int x_1_1_2; unsigned int x_1_1_3; int x_1_1_4; } x1; struct { int x_2_1_1; int x_2_1_2; unsigned int x_2_1_3; int x_2_1_4; } x2; })arg2 toChapters:(id)arg3 fromIndex:(unsigned int)arg4;
+- (unsigned int)_addChapterMetadataItem:(id)arg1 timeRange:(struct { struct { long long x_1_1_1; int x_1_1_2; unsigned int x_1_1_3; long long x_1_1_4; } x1; struct { long long x_2_1_1; int x_2_1_2; unsigned int x_2_1_3; long long x_2_1_4; } x2; })arg2 toChapters:(id)arg3 fromIndex:(unsigned int)arg4;
 - (id)_assetInspector;
 - (id)_assetInspectorLoader;
 - (id)_availableCanonicalizedChapterLanguages;
@@ -81,7 +81,7 @@
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
 - (id)creationDate;
 - (void)dealloc;
-- (struct { int x1; int x2; unsigned int x3; int x4; })duration;
+- (struct { long long x1; int x2; unsigned int x3; long long x4; })duration;
 - (BOOL)hasProtectedContent;
 - (unsigned int)hash;
 - (id)init;
@@ -103,12 +103,12 @@
 - (id)mediaSelectionGroupForPropertyList:(id)arg1 mediaSelectionOption:(id*)arg2;
 - (id)metadata;
 - (id)metadataForFormat:(id)arg1;
-- (struct CGSize { double x1; double x2; })naturalSize;
+- (struct CGSize { float x1; float x2; })naturalSize;
 - (int)naturalTimeScale;
 - (id)preferredMediaSelection;
 - (float)preferredRate;
 - (float)preferredSoundCheckVolumeNormalization;
-- (struct CGAffineTransform { double x1; double x2; double x3; double x4; double x5; double x6; })preferredTransform;
+- (struct CGAffineTransform { float x1; float x2; float x3; float x4; float x5; float x6; })preferredTransform;
 - (float)preferredVolume;
 - (id)propertyListForProxy;
 - (BOOL)providesPreciseDurationAndTiming;
@@ -137,7 +137,7 @@
 
 - (void)_pu_setCachedDuration:(id)arg1;
 - (id)pu_cachedDuration;
-- (struct { int x1; int x2; unsigned int x3; int x4; })pu_duration;
+- (struct { long long x1; int x2; unsigned int x3; long long x4; })pu_duration;
 - (void)pu_loadDurationWithCompletionHandler:(id /* block */)arg1;
 
 // Image: /System/Library/PrivateFrameworks/NotesShared.framework/NotesShared
@@ -149,7 +149,7 @@
 
 - (float)is_cropFactor;
 - (id)is_valueForMetadataIdentifier:(id)arg1;
-- (struct CGSize { double x1; double x2; })is_videoSize;
+- (struct CGSize { float x1; float x2; })is_videoSize;
 
 // Image: /System/Library/PrivateFrameworks/VideoProcessing.framework/VideoProcessing
 
@@ -166,6 +166,6 @@
 
 + (id)keyPathsForValuesAffectingNaturalSizeWithPreferredTransforms;
 
-- (struct CGSize { double x1; double x2; })naturalSizeWithPreferredTransforms;
+- (struct CGSize { float x1; float x2; })naturalSizeWithPreferredTransforms;
 
 @end
