+++ Frameworks/AVFoundation.framework/AVAssetMakeReadyForInspectionLoader.h
@@ -36,7 +36,7 @@
 - (void)cancelLoading;
 - (id)copyWithZone:(struct _NSZone { }*)arg1;
 - (void)dealloc;
-- (struct { int x1; int x2; unsigned int x3; int x4; })duration;
+- (struct { long long x1; int x2; unsigned int x3; long long x4; })duration;
 - (id)figChapterGroupInfo;
 - (void)finalize;
 - (BOOL)hasProtectedContent;
