+++ Frameworks/AVFoundation.framework/AVAssetReaderInternal.h
@@ -14,16 +14,16 @@
     NSObject<OS_dispatch_queue> * statusReadWriteQueue;
     struct { 
         struct { 
-            int value; 
+            long long value; 
             int timescale; 
             unsigned int flags; 
-            int epoch; 
+            long long epoch; 
         } start; 
         struct { 
-            int value; 
+            long long value; 
             int timescale; 
             unsigned int flags; 
-            int epoch; 
+            long long epoch; 
         } duration; 
     }  timeRange;
     AVWeakReference * weakReference;
