+++ Frameworks/AVFoundation.framework/AVAssetResourceLoadingContentInformationRequest.h
@@ -7,11 +7,11 @@
 }
 
 @property (getter=isByteRangeAccessSupported, nonatomic) BOOL byteRangeAccessSupported;
-@property (nonatomic) int contentLength;
+@property (nonatomic) long long contentLength;
 @property (nonatomic, copy) NSString *contentType;
 @property (nonatomic, copy) NSDate *renewalDate;
 
-- (int)contentLength;
+- (long long)contentLength;
 - (id)contentType;
 - (void)dealloc;
 - (id)description;
@@ -23,7 +23,7 @@
 - (id)propertyList;
 - (id)renewalDate;
 - (void)setByteRangeAccessSupported:(BOOL)arg1;
-- (void)setContentLength:(int)arg1;
+- (void)setContentLength:(long long)arg1;
 - (void)setContentType:(id)arg1;
 - (void)setDiskCachingPermitted:(BOOL)arg1;
 - (void)setRenewalDate:(id)arg1;
