+++ Frameworks/AVFoundation.framework/AVAssetSynchronousInspectorLoader.h
@@ -8,7 +8,7 @@
 
 - (id)assetInspector;
 - (void)dealloc;
-- (struct { int x1; int x2; unsigned int x3; int x4; })duration;
+- (struct { long long x1; int x2; unsigned int x3; long long x4; })duration;
 - (unsigned int)hash;
 - (id)initWithAssetInspector:(id)arg1;
 - (BOOL)isCompatibleWithSavedPhotosAlbum;
